move namespace logic to sparate file, fix bug related to the q:template
