+++ packages/qwik/src/core/v2/client/vnode-diff.ts
@@ -19,15 +19,12 @@ import {
   ELEMENT_KEY,
   ELEMENT_PROPS,
   ELEMENT_SEQ,
-  HTML_NS,
-  MATH_NS,
   OnRenderProp,
   QDefaultSlot,
   QSlot,
   QSlotParent,
   QStyle,
   QTemplate,
-  SVG_NS,
 } from '../../util/markers';
 import { isPromise } from '../../util/promises';
 import { type ValueOrPromise } from '../../util/types';
@@ -58,9 +55,6 @@ import {
   type VirtualVNode,
 } from './types';
 import {
-  isForeignObjectElement,
-  isMathElement,
-  isSvgElement,
   mapApp_findIndx,
   mapArray_set,
   vnode_ensureElementInflated,
@@ -77,7 +71,6 @@ import {
   vnode_getText,
   vnode_getType,
   vnode_insertBefore,
-  vnode_isDefaultNamespace,
   vnode_isElementVNode,
   vnode_isTextVNode,
   vnode_isVNode,
@@ -93,6 +86,7 @@ import {
   vnode_truncate,
   type VNodeJournal,
 } from './vnode';
+import { getNewElementNamespaceData } from './vnode-namespace';
 
 export type ComponentQueue = Array<VNode>;
 
@@ -632,50 +626,14 @@ export const vnode_diff = (
 
   function createElementWithNamespace(tag: string): Element {
     const domParentVNode = vnode_getDomParentVNode(vParent);
+    const { elementNamespace, elementNamespaceFlag } = getNewElementNamespaceData(
+      domParentVNode,
+      tag
+    );
 
-    const parentIsDefaultNamespace = domParentVNode
-      ? vnode_getElementName(domParentVNode) && vnode_isDefaultNamespace(domParentVNode)
-      : true;
-    const parentIsForeignObject = !parentIsDefaultNamespace
-      ? isForeignObjectElement(vnode_getElementName(domParentVNode!))
-      : false;
-
-    let elementNamespace = HTML_NS;
-    let elementNamespaceFlag = VNodeFlags.NS_html;
-
-    if (isSvgElement(tag)) {
-      elementNamespace = SVG_NS;
-      elementNamespaceFlag = VNodeFlags.NS_svg;
-    } else if (isMathElement(tag)) {
-      elementNamespace = MATH_NS;
-      elementNamespaceFlag = VNodeFlags.NS_math;
-    } else if (domParentVNode && !parentIsForeignObject) {
-      const isParentSvg = (domParentVNode[VNodeProps.flags] & VNodeFlags.NS_svg) !== 0;
-      const isParentMath = (domParentVNode[VNodeProps.flags] & VNodeFlags.NS_math) !== 0;
-
-      elementNamespace = isParentSvg ? SVG_NS : isParentMath ? MATH_NS : HTML_NS;
-      elementNamespaceFlag = domParentVNode[VNodeProps.flags] & VNodeFlags.NAMESPACE_MASK;
-    } else {
-      elementNamespace = HTML_NS;
-      elementNamespaceFlag = VNodeFlags.NS_html;
-    }
-
-    const newChildIsDefaultNamespace = elementNamespace === HTML_NS;
-
-    let element: Element;
-    if (parentIsDefaultNamespace && newChildIsDefaultNamespace) {
-      element = container.document.createElement(tag);
-      vNewNode = vnode_newElement(element, tag);
-    } else {
-      element = container.document.createElementNS(elementNamespace, tag);
-      vNewNode = vnode_newElement(element, tag);
-
-      if (newChildIsDefaultNamespace) {
-        vNewNode[VNodeProps.flags] |= domParentVNode![VNodeProps.flags] & VNodeFlags.NAMESPACE_MASK;
-      } else {
-        vNewNode[VNodeProps.flags] |= elementNamespaceFlag;
-      }
-    }
+    const element = container.document.createElementNS(elementNamespace, tag);
+    vNewNode = vnode_newElement(element, tag);
+    vNewNode[VNodeProps.flags] |= elementNamespaceFlag;
     return element;
   }
